Fix bug, refactor code
