+++ src/main/java/com/moviesandchill/portalbackendservice/controller/UserController.java
@@ -1,9 +1,9 @@
 package com.moviesandchill.portalbackendservice.controller;
 
-import com.moviesandchill.portalbackendservice.dto.password.UpdatePasswordDto;
 import com.moviesandchill.portalbackendservice.dto.user.FullUserDto;
 import com.moviesandchill.portalbackendservice.dto.user.NewUserDto;
 import com.moviesandchill.portalbackendservice.dto.user.UserDto;
+import com.moviesandchill.portalbackendservice.dto.user.password.UpdatePasswordDto;
 import com.moviesandchill.portalbackendservice.mapper.CommonMapper;
 import com.moviesandchill.portalbackendservice.service.user.UserService;
 import org.springframework.http.ResponseEntity;
