+++ src/main/java/com/moviesandchill/portalbackendservice/dto/film/agelimit/AgeLimitDto.java
@@ -1,4 +1,4 @@
-package com.moviesandchill.portalbackendservice.dto.film;
+package com.moviesandchill.portalbackendservice.dto.film.agelimit;
 
 import lombok.*;
