+++ src/main/java/com/moviesandchill/portalbackendservice/dto/film/film/FilmDto.java
@@ -1,4 +1,4 @@
-package com.moviesandchill.portalbackendservice.dto.film;
+package com.moviesandchill.portalbackendservice.dto.film.film;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
