+++ src/main/java/com/moviesandchill/portalbackendservice/dto/film/gender/GenderDto.java
@@ -1,14 +1,18 @@
-package com.moviesandchill.portalbackendservice.dto.globalrole;
+package com.moviesandchill.portalbackendservice.dto.film.gender;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class UpdateGlobalRoleDto {
-    private String name;
+public class GenderDto {
+
+    private Long id_gender;
+
+    private String gender_title;
 }
