+++ src/main/java/com/moviesandchill/portalbackendservice/dto/film/review/ReviewDto.java
@@ -1,4 +1,4 @@
-package com.moviesandchill.portalbackendservice.dto.film;
+package com.moviesandchill.portalbackendservice.dto.film.review;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
