+++ src/main/java/com/moviesandchill/portalbackendservice/dto/film/screenshot/ScreenshotDto.java
@@ -1,15 +1,18 @@
-package com.moviesandchill.portalbackendservice.dto.globalrole;
+package com.moviesandchill.portalbackendservice.dto.film.screenshot;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class GlobalRoleDto {
-    private Long globalRoleId;
-    private String name;
+public class ScreenshotDto {
+
+    private Long id_screenshot;
+
+    private String screenshot;
 }
