+++ src/main/java/com/moviesandchill/portalbackendservice/dto/film/staffrole/StaffRoleDto.java
@@ -1,14 +1,18 @@
-package com.moviesandchill.portalbackendservice.dto.globalrole;
+package com.moviesandchill.portalbackendservice.dto.film.staffrole;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class NewGlobalRoleDto {
-    private String name;
+public class StaffRoleDto {
+
+    private Long id_staff_role;
+
+    private String role_title;
 }
