--- src/main/java/com/moviesandchill/portalbackendservice/dto/login/LoginRequestDto.java
@@ -1,16 +1,0 @@
-package com.moviesandchill.portalbackendservice.dto.login;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class LoginRequestDto {
-    private String login;
-
-    private String password;
-}
