--- src/main/java/com/moviesandchill/portalbackendservice/dto/password/UpdatePasswordDto.java
@@ -1,15 +1,0 @@
-package com.moviesandchill.portalbackendservice.dto.password;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class UpdatePasswordDto {
-    private String oldPassword;
-    private String newPassword;
-}
