+++ src/main/java/com/moviesandchill/portalbackendservice/dto/user/achievement/AchievementDto.java
@@ -1,0 +1,20 @@
+package com.moviesandchill.portalbackendservice.dto.user.achievement;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class AchievementDto {
+    private Long achievementId;
+
+    private String name;
+
+    private String logoUrl;
+
+    private String description;
+}
