+++ src/main/java/com/moviesandchill/portalbackendservice/dto/user/globalrole/NewGlobalRoleDto.java
@@ -1,18 +1,14 @@
-package com.moviesandchill.portalbackendservice.dto.film;
+package com.moviesandchill.portalbackendservice.dto.user.globalrole;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class CountryDto {
-
-    private Long id_country;
-
-    private String country_name;
+public class NewGlobalRoleDto {
+    private String name;
 }
