+++ src/main/java/com/moviesandchill/portalbackendservice/dto/user/globalrole/UpdateGlobalRoleDto.java
@@ -1,18 +1,14 @@
-package com.moviesandchill.portalbackendservice.dto.film;
+package com.moviesandchill.portalbackendservice.dto.user.globalrole;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class GenreDto {
-
-    private Long id_genre;
-
-    private String genre_title;
+public class UpdateGlobalRoleDto {
+    private String name;
 }
