+++ src/main/java/com/moviesandchill/portalbackendservice/dto/user/password/UpdatePasswordDto.java
@@ -1,18 +1,15 @@
-package com.moviesandchill.portalbackendservice.dto.film;
+package com.moviesandchill.portalbackendservice.dto.user.password;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class GenderDto {
-
-    private Long id_gender;
-
-    private String gender_title;
+public class UpdatePasswordDto {
+    private String oldPassword;
+    private String newPassword;
 }
