+++ src/main/java/com/moviesandchill/portalbackendservice/service/AuthService.java
@@ -1,8 +1,8 @@
 package com.moviesandchill.portalbackendservice.service;
 
-import com.moviesandchill.portalbackendservice.dto.login.LoginRequestDto;
 import com.moviesandchill.portalbackendservice.dto.user.NewUserDto;
 import com.moviesandchill.portalbackendservice.dto.user.UserDto;
+import com.moviesandchill.portalbackendservice.dto.user.login.LoginRequestDto;
 
 import java.util.Optional;
