+++ src/main/java/com/moviesandchill/portalbackendservice/service/film/FilmService.java
@@ -1,6 +1,10 @@
 package com.moviesandchill.portalbackendservice.service.film;
 
-import com.moviesandchill.portalbackendservice.dto.film.*;
+import com.moviesandchill.portalbackendservice.dto.film.film.FilmDto;
+import com.moviesandchill.portalbackendservice.dto.film.genre.GenreDto;
+import com.moviesandchill.portalbackendservice.dto.film.review.ReviewDto;
+import com.moviesandchill.portalbackendservice.dto.film.screenshot.ScreenshotDto;
+import com.moviesandchill.portalbackendservice.dto.film.staff.StaffDto;
 
 import java.util.List;
 import java.util.Optional;
