+++ src/main/java/com/moviesandchill/portalbackendservice/service/impl/user/UserAchievementServiceImpl.java
@@ -1,6 +1,6 @@
 package com.moviesandchill.portalbackendservice.service.impl.user;
 
-import com.moviesandchill.portalbackendservice.dto.achievement.AchievementDto;
+import com.moviesandchill.portalbackendservice.dto.user.achievement.AchievementDto;
 import com.moviesandchill.portalbackendservice.mapper.CommonMapper;
 import com.moviesandchill.portalbackendservice.service.user.UserAchievementService;
 import lombok.extern.slf4j.Slf4j;
