+++ src/main/java/com/moviesandchill/portalbackendservice/service/user/UserGlobalRoleService.java
@@ -1,6 +1,6 @@
 package com.moviesandchill.portalbackendservice.service.user;
 
-import com.moviesandchill.portalbackendservice.dto.globalrole.GlobalRoleDto;
+import com.moviesandchill.portalbackendservice.dto.user.globalrole.GlobalRoleDto;
 
 import java.util.List;
